Backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
httpx==0.25.0

# LangGraph and LLM utilities
langgraph==0.1.11
litellm==1.9.1
langchain==0.0.335
langchain-community==0.0.13

# Vector database clients
qdrant-client==1.6.4
neo4j==5.13.0

# Data processing
numpy==1.26.1
pandas==2.1.2

# Async support
asyncio==3.4.3
aiohttp==3.8.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
